fix:リクエスト取得用構造体を変更
リクエスト取得用構造体の変更
